refactor: collapse QA report shim (#1848)
Remove the legacy polylogue.showcase.qa_report aggregator and route callers to the owned QA payload, Markdown, and summary modules directly. The validation-lab runner, report persistence path, and generated report artifacts remain intact.\n\nUpdate the #1848 static rendering classification ledger and topology generated surfaces to record the removed module.\n\nVerification:\n- devtools test tests/unit/showcase/test_report.py tests/unit/showcase/test_operator_reporting_runtime.py tests/unit/cli/test_deterministic_output.py -k 'ShowcaseReport or qa or QA or FrozenClockShowcaseReport or save_qa_reports or format_qa_summary or report_common or legacy_qa_report' tests/unit/architecture/test_static_rendering_prune.py: 18 passed, 97 deselected\n- devtools verify --quick: exit_code 0
