+++ docs/topology-status.md
@@ -28,10 +28,10 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 693
+- **Stable** (no move scoped): 692
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 16
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 829
-- **Realized polylogue/**/*.py**: 829 files declared
+- **Total declared**: 828
+- **Realized polylogue/**/*.py**: 828 files declared
 <!-- END GENERATED: topology-status -->
